Fixing packages and starting to create the execute method

+++ user-service/src/main/java/com/test/userservice/dto/UserResponse.java
@@ -2,7 +2,5 @@ package com.test.userservice.dto;
 
 import com.test.userservice.enums.UserStatus;
 
-import java.math.BigDecimal;
-
-public record UserResponse(Long id, String username, String password, UserStatus status, BigDecimal balance) {
+public record UserResponse(Long id, String username, String password, UserStatus status, Double balance) {
 }
